requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.9
eventlet==0.33.3
python-engineio==4.8.0
python-socketio==5.10.0
Werkzeug==2.3.7
Flask-WTF